Added basic HTML layout, random images, + sidebar
First commit for this project. Created the generate button, the output div with random image,
and the sidebar that slides in from the right.
Buttons for opening and closing the sidebar function. 'Generate' button generates a random
number and image between 0 and 14.
Uploaded 15 images to be backgrounds for the generated number. Images all have size
300x300 px and were made by yours truly.